Aigerim (#89)
If you can avoid docx, because I can not see them on github. 
Docx is a bad format in term of efficiency.

You can faster write on markdown I would say. 

If you continue to work with docx it is fine, but I would not recommend it for programming works.